requirements.txt
aiogram==2.25.*
python-dotenv
peewee